# api/docs/swagger.yaml
definitions:
...
  models.CreateTarif:
    properties:
      amount_for_card:
        type: number
      amount_for_cash:
        type: number
      created_at:
        type: string
      name:
        type: string
      tarif_type:
        type: string
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      status_code:
        type: integer
    type: object
...
  models.Tarif:
    properties:
      amount_for_card:
        type: number
      amount_for_cash:
        type: number
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      tarif_type:
        type: string
      updated_at:
        type: string
    type: object
  models.TarifsResponse:
    properties:
      count:
        type: integer
      tarifs:
        items:
          $ref: '#/definitions/models.Tarif'
        type: array
    type: object
...
  models.UpdateTarif:
    properties:
      amount_for_card:
        type: number
      amount_for_cash:
        type: number
      id:
        type: string
      name:
        type: string
      tarif_type:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /tarif:
    post:
      consumes:
      - application/json
      description: create a new tarif
      parameters:
      - description: tarif
        in: body
        name: tarif
        schema:
          $ref: '#/definitions/models.CreateTarif'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tarif'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create a new tarif
      tags:
      - tarif
  /tarif/{id}:
    delete:
      consumes:
      - application/json
      description: delete tarif
      parameters:
      - description: tarif_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete tarif
      tags:
      - tarif
    get:
      consumes:
      - application/json
      description: get tarif by id
      parameters:
      - description: tarif_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tarif'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get tarif by id
      tags:
      - tarif
    put:
      consumes:
      - application/json
      description: update tarif
      parameters:
      - description: tarif_id
        in: path
        name: id
        required: true
        type: string
      - description: tarif
        in: body
        name: tarif
        schema:
          $ref: '#/definitions/models.UpdateTarif'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tarif'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update tarif
      tags:
      - tarif
  /tarifs:
    get:
      consumes:
      - application/json
      description: get tarifs list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TarifsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get tarifs list
      tags:
      - tarifs
...
swagger: "2.0"